scripts/extraction_data/requirements.txt
PyMuPDF>=1.24
pdfplumber>=0.11
pandas>=2.2
Pillow>=10.0
pytesseract>=0.3.10
